requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
psycopg2==2.9.3
pydantic==2.10.2
python-dotenv==1.0.1
requests==2.32.3
